feat: update user controller

+++ src/user/user.controller.ts
@@ -6,10 +6,14 @@ import {
   Param,
   Delete,
   Put,
+  HttpException,
+  HttpStatus,
+  HttpCode,
 } from '@nestjs/common';
 import { UserService } from './user.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdatePasswordDto } from './dto/update-password.dto';
+import { isUUID } from 'class-validator';
 
 @Controller('user')
 export class UserController {
@@ -22,7 +26,20 @@ export class UserController {
 
   @Get(':id')
   findOne(@Param('id') id: string) {
-    return this.userService.findOne(id);
+    if (!isUUID(id, 4)) {
+      throw new HttpException(
+        'User ID is invalid (not uuid)',
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+
+    const user = this.userService.findOne(id);
+
+    if (!user) {
+      throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+    }
+
+    return user;
   }
 
   @Post()
@@ -35,11 +52,40 @@ export class UserController {
     @Param('id') id: string,
     @Body() updatePasswordDto: UpdatePasswordDto,
   ) {
-    return this.userService.update(id, updatePasswordDto);
+    if (!isUUID(id, 4)) {
+      throw new HttpException(
+        'User ID is invalid (not uuid)',
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+
+    const user = this.userService.findOne(id);
+
+    if (!user) {
+      throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+    }
+
+    if (user.password !== updatePasswordDto.oldPassword) {
+      throw new HttpException('Old password is wrong', HttpStatus.FORBIDDEN);
+    }
+
+    return this.userService.update(id, updatePasswordDto.newPassword);
   }
 
   @Delete(':id')
+  @HttpCode(204)
   remove(@Param('id') id: string) {
-    return this.userService.remove(id);
+    if (!isUUID(id, 4)) {
+      throw new HttpException(
+        'User ID is invalid (not uuid)',
+        HttpStatus.BAD_REQUEST,
+      );
+    }
+
+    const user = this.userService.remove(id);
+
+    if (user === false) {
+      throw new HttpException('User not found', HttpStatus.NOT_FOUND);
+    }
   }
 }
